Narrow six returns that ty 0.0.69 calls unsound
The pending dependency bump takes ty from 0.0.65 to 0.0.69, which adds
`unsound-return-statement`: a value inferred as Any (or a union containing it)
returned from a function that declares a concrete type. Six sites trip it and
the bump's CI fails on all three Python versions because of them.

None is in `src/`. Five come from arithmetic over values the checker cannot see
into, and one from `tomllib`, which is Any by construction. The fix is the
narrowing the diagnostic asks for rather than a rule exemption: `float(...)` and
`int(...)` assert at runtime what the annotation already promised, and the two
`_agreement` calls get annotated locals because the script they come from is
imported dynamically, so every attribute off it is Any.

Verified under both versions -- 0.0.65 as pinned on main and 0.0.69 as the bump
would install -- with `uv run ty check`, `ruff format --check`, and `ruff check`
run exactly as ci.yml runs them, plus the tests for the three test files touched.
No lockfile change here: pinning the new ty is the bump's job, not this one's.

+++ scripts/measure_correlated_fleets.py
@@ -88,10 +88,12 @@ def exact_wrong_majority(rate: float, *, schools: int, fleet: int = FLEET) -> fl
     """
     per_school = fleet // schools
     needed = fleet // 2 + 1
-    return sum(
-        comb(schools, j) * rate**j * (1 - rate) ** (schools - j)
-        for j in range(schools + 1)
-        if j * per_school >= needed
+    return float(
+        sum(
+            comb(schools, j) * rate**j * (1 - rate) ** (schools - j)
+            for j in range(schools + 1)
+            if j * per_school >= needed
+        )
     )
